refactor: move remaining private methods in main class to a mixin
